+++ MapleStory_OEH/MapleCore/Level_BeginnersTown2.cpp
@@ -188,7 +188,7 @@ void Level_BeginnersTown2::UnLoadResources()
 
 			std::string FileFullPath = File[i].GetFullPath();
 			std::string FileName = "";
-			int Count = 0;
+			size_t Count = 0;
 
 			for (Count = FileFullPath.size(); Count > 0; Count--)
 			{
@@ -199,7 +199,7 @@ void Level_BeginnersTown2::UnLoadResources()
 				}
 			}
 
-			for (int j = Count + 1; j < FileFullPath.size(); j++)
+			for (size_t j = Count + 1; j < FileFullPath.size(); j++)
 			{
 				FileName.push_back(FileFullPath[j]);
 			}
